requirements.txt
# used packages

pip==20.3.3
ipython==7.19.0
pillow==8.0.1
numpy==1.19.4
torch==1.7.1
torchvision==0.8.2
google-cloud-stroage==1.35.0
pandas==1.1.5
matplotlib==3.3.3
torch-summary==1.4.3
